MAF column candidate names not in upper  case when matching on upper case normalized names
I want to start by saying, thank you for providing this very helpful package with your paper.

When matching on the column name for MAF, the vector holding the data on line 99 is not in upper case, but `hold_names` has been converted to upper case in line 28.

This would mean that for many cases, e.g. CEUaf, the MAF column would not be identified.

I propose to wrap `toupper` around the vector containing the data, since that is the minimal change for this to work correctly.

A better solution would be to factor out all of these "potential names" vectors out of the function. Then one could replace the vectors with a function call like `maf_names()` or another named vector, e.g. `maf_names`.  This would make the code more clear and concise, plus it would encapsulate the data defined by these vectors in a function or some other entity defined in a single place.